+++ my-cdk-project/lambda-functions/getProductsById.ts
@@ -1,0 +1,27 @@
+import { APIGatewayProxyEvent, APIGatewayProxyHandler, APIGatewayProxyResult } from "aws-lambda";
+import { products } from "./mockData";
+
+export const handler: APIGatewayProxyHandler = async (
+  event: APIGatewayProxyEvent,
+): Promise<APIGatewayProxyResult> => {  
+
+  const productId = event.pathParameters?.productId;
+  const product = products.find(p => p.id === productId);
+  if (product) {
+    return {
+      statusCode: 200,
+      headers: {
+        "Access-Control-Allow-Origin": "*",
+        "Access-Control-Allow-Methods": "GET",
+        "Access-Control-Allow-Headers": "Content-Type",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(product),
+    };
+  } else {
+    return {
+      statusCode: 404,
+      body: JSON.stringify({ error: 'Product not found' }),
+    };
+  }
+};
