+++ my-cdk-project/lambda-functions/mockData.ts
@@ -1,0 +1,7 @@
+import { IProduct } from "./productInterface";
+
+export const products: IProduct[] = [
+  { id: '1', title: 'Product 1', price: 100, description: 'super', count: 5 },
+      { id: '2', title: 'Product 2', price: 150, description: 'bad', count: 4 },
+      { id: '3', title: 'Product 3', price: 200, description: 'norm', count: 5 }
+];
