+++ my-cdk-project/lambda-functions/productInterface.ts
