+++ my-cdk-project/package.json
@@ -6,6 +6,7 @@
   },
   "scripts": {
     "build": "tsc",
+    "cdk:deploy": "npm run build && cdk deploy",
     "watch": "tsc -w",
     "test": "jest",
     "cdk": "cdk"
